refactor(runtime): extract runtime test helpers
