+++ ui/src/lib/runtime/PatchRuntime.test-helpers.ts
@@ -1,0 +1,152 @@
+import type { Node } from '@xyflow/svelte';
+import { vi } from 'vitest';
+import type { MessageContext } from '$lib/messages/MessageContext';
+import { ObjectContext } from '$lib/objects/v2/ObjectContext';
+import type { TextObjectV2 } from '$lib/objects/v2/interfaces/text-objects';
+import type { ObjectInlet, ObjectOutlet } from '$lib/objects/v2/object-metadata';
+
+import type { EditorRuntime } from './EditorRuntimeReconciler';
+import type { PatchRuntimeObjectService } from './PatchMessageRuntime';
+
+export const TEST_OBJECT_TYPE = 'patch-runtime-test';
+
+export class PatchRuntimeTestObject implements TextObjectV2 {
+  static type = TEST_OBJECT_TYPE;
+  static inlets: ObjectInlet[] = [{ name: 'value', type: 'any' as const }];
+  static outlets: ObjectOutlet[] = [{ name: 'out', type: 'any' as const }];
+
+  static createdRawParams: unknown[][] = [];
+  static destroyedNodeIds: string[] = [];
+  static createGate: Promise<void> | null = null;
+  static normalizeParamOnCreate = false;
+  static dynamicInlets: ObjectInlet[] | null = null;
+  static dynamicOutlets: ObjectOutlet[] | null = null;
+
+  constructor(
+    readonly nodeId: string,
+    readonly context: ObjectContext
+  ) {}
+
+  async create(rawParams: unknown[]) {
+    await PatchRuntimeTestObject.createGate;
+
+    PatchRuntimeTestObject.createdRawParams.push(rawParams);
+
+    if (PatchRuntimeTestObject.normalizeParamOnCreate) {
+      this.context.setParam(0, 'normalized');
+    }
+  }
+
+  destroy() {
+    PatchRuntimeTestObject.destroyedNodeIds.push(this.nodeId);
+  }
+
+  getInlets() {
+    return PatchRuntimeTestObject.dynamicInlets ?? PatchRuntimeTestObject.inlets;
+  }
+
+  getOutlets() {
+    return PatchRuntimeTestObject.dynamicOutlets ?? PatchRuntimeTestObject.outlets;
+  }
+}
+
+export function resetPatchRuntimeTestObject(): void {
+  PatchRuntimeTestObject.createdRawParams = [];
+  PatchRuntimeTestObject.destroyedNodeIds = [];
+  PatchRuntimeTestObject.createGate = null;
+  PatchRuntimeTestObject.normalizeParamOnCreate = false;
+  PatchRuntimeTestObject.dynamicInlets = null;
+  PatchRuntimeTestObject.dynamicOutlets = null;
+}
+
+export class FakeObjectService implements PatchRuntimeObjectService {
+  private objectsById = new Map<string, TextObjectV2>();
+
+  isObjectInRegistry(objectType: string): boolean {
+    return objectType === TEST_OBJECT_TYPE;
+  }
+
+  async createObject(
+    nodeId: string,
+    objectType: string,
+    messageContext: MessageContext,
+    params: unknown[] = [],
+    rawParams: string[] = []
+  ): Promise<TextObjectV2 | null> {
+    if (!this.isObjectInRegistry(objectType)) return null;
+
+    const context = new ObjectContext(nodeId, messageContext, PatchRuntimeTestObject.inlets);
+    context.initParams(params);
+
+    const object = new PatchRuntimeTestObject(nodeId, context);
+    this.objectsById.set(nodeId, object);
+
+    await object.create(rawParams);
+
+    return object;
+  }
+
+  removeObjectById(nodeId: string): void {
+    const object = this.objectsById.get(nodeId);
+    if (!object) return;
+
+    object.destroy?.();
+    object.context.destroy();
+
+    this.objectsById.delete(nodeId);
+  }
+
+  getObjectById(nodeId: string): TextObjectV2 | null {
+    return this.objectsById.get(nodeId) ?? null;
+  }
+}
+
+export class FakeAudioService {
+  removeNodeById = vi.fn();
+  createNode = vi.fn();
+  updateEdges = vi.fn();
+  send = vi.fn();
+
+  audioNode = {
+    nodeId: 'object-audio-runtime-test',
+    audioNode: null
+  };
+
+  getNodeById = vi.fn(() => this.audioNode);
+}
+
+export class FakeEventBus {
+  listeners = new Map<string, Array<(event: never) => void>>();
+
+  addEventListener = vi.fn((type: string, listener: (event: never) => void) => {
+    this.listeners.set(type, [...(this.listeners.get(type) ?? []), listener]);
+  });
+
+  removeEventListener = vi.fn((type: string, listener: (event: never) => void) => {
+    this.listeners.set(
+      type,
+      (this.listeners.get(type) ?? []).filter((candidate) => candidate !== listener)
+    );
+  });
+
+  dispatch(event: { type: string } & Record<string, unknown>) {
+    for (const listener of this.listeners.get(event.type) ?? []) {
+      listener(event as never);
+    }
+  }
+}
+
+export const objectNode = (id: string, data: Record<string, unknown>): Node => ({
+  id,
+  type: 'object',
+  position: { x: 0, y: 0 },
+  data
+});
+
+export const createFakeEditorRuntime = (overrides: Partial<EditorRuntime> = {}) => ({
+  isObjectInRegistry: vi.fn((objectType: string) => objectType === TEST_OBJECT_TYPE),
+  createObject: vi.fn(),
+  updateObject: vi.fn(),
+  destroyObject: vi.fn(),
+  ...overrides
+});
